refactor: address 13 code review findings
- F1: Delete duplicate format_error_with_backup_hint/build_error_output
  from cmd/tx.rs, import from crate::tx
- F2: Extract resolve_target_paths helper in ast.rs (6 identical blocks)
- F3: Extract apply_or_preview helper in ast.rs (2 duplicated blocks)
- F4: Extract ~400 lines of YAML text-splice functions from ops/doc.rs
  into ops/doc/yaml_splice.rs submodule
- F5: Add DocAction::is_write()/file_path() methods in cmd/doc.rs,
  simplify run() function
- F6: Remove identity match in write.rs (normalize_eol map)
- F7: Extract ~1,867 lines of tests from api/mod.rs into api/tests.rs
- F8: Replace 13 manual JSON/JSONL dispatch blocks in ast.rs with
  global.emit_json() calls (also fixes 2 sites that always used
  to_string_pretty, ignoring --jsonl)
- F9: Make collect_file_paths delegate to collect_file_paths_with_ignores
- F10: Extract shared run_lifecycle_steps helper in tx.rs
- F11: Add md_edit helper in api/md.rs, refactor 5 functions to use it
- F12: Split operation_schemas() into per-tier functions in schema.rs
- F13: Remove duplicate #[cfg(feature = "cli")] attribute in files.rs
